Wait for agent departures
If an agent is uninstalled from the `telepresence-agents` configmap, and
a `telepresence intercept` is made shortly after, then it's possible
that the traffic-manager awaits the agent's arrival before the removed
agent departs. This commit ensures that this situation is detected,
so the await for arrival first ensures that all agents affected by the
removal have departed before a new agent is added.